updating empty_observation to ObservationArray
