+++ rubin_scheduler/scheduler/surveys/pointings_survey.py
@@ -21,7 +21,7 @@ class PointingsSurvey(BaseSurvey):
     ----------
     observations : `np.array`
         An array of observations, from e.g.,
-        rubin_scheduler.scheduler.utils.empty_observation
+        rubin_scheduler.scheduler.utils.ObservationArray
         expect "RA", "dec", and "note" to be filled, other columns ignored.
     gap_min : `float`
         The minimum gap to force between observations of the same
